regrouping components (moving sidebar components into AdvancedPortfolioManager folder)
