+++ static-webapps/slate-portfolio-manager/src/views/AdvancedPortfolioManager/AdvancedPortfolioSidebar/SkillDemos.vue
