+++ static-webapps/slate-portfolio-manager/src/views/AdvancedPortfolioManager/AdvancedPortfolioSidebar/index.js
@@ -1,0 +1,3 @@
+import AdvancedPortfolioSidebar from './AdvancedPortfolioSidebar.vue'
+
+export default AdvancedPortfolioSidebar
