+++ static-webapps/slate-portfolio-manager/src/views/AdvancedPortfolioManager/index.js
@@ -1,0 +1,3 @@
+import AdvancedPortfolioManager from './AdvancedPortfolioManager.vue'
+
+export default AdvancedPortfolioManager
