+++ static-webapps/slate-portfolio-manager/src/views/index.js
@@ -1,0 +1,5 @@
+import AdvancedPortfolioManager from './AdvancedPortfolioManager'
+
+export default {
+  AdvancedPortfolioManager,
+}
